Đã xong kéo thả tháp, nhưng vị trí đặt còn chưa chọn lọc, có thể đặt tháp mọi nơi trên bản đồ, đạn chưa bắn ra khi đặt tháp. Tower image cần giúp xóa phông
