+++ src/sample/Tower.java
@@ -1,6 +1,13 @@
 package sample;
 
+import javafx.event.EventHandler;
+import javafx.scene.Cursor;
+import javafx.scene.Scene;
 import javafx.scene.effect.Light;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.input.*;
+import javafx.scene.layout.HBox;
 
 
 public abstract class Tower extends GameTile{
@@ -14,6 +21,10 @@ public abstract class Tower extends GameTile{
     abstract public void shoot();
 
 //    abstract public
+    @Override
+    public void loadImage(String path) {
+    this.image = new Image(path + ".png", 50, 50, true, true);
+}
 
 
 }
